Add only_current option for request_company_report
BambooHR company reports by default will return None for some
fields for employees who's start date is in the future, such
as the supervisorEId. This makes it impossible to use reports
for pre-hire activities.

BambooHR support noted the undocumented "onlyCurrent" parameter
allows you to change this behaviour.

This change allows you to turn this on or off. It defaults to
the more sane version where all data is returned, even for future
employees.